Warning fixes from bad merge.

+++ JDFlipNumberView/JDFlipNumberViewImageFactorySKZ.m
@@ -110,7 +110,7 @@ static JDFlipNumberViewImageFactorySKZ *sharedInstance;
 	// create bottom and top images
     for (NSInteger j=0; j<10; j++) {
         for (int i=0; i<2; i++) {
-            NSString *imageName = [NSString stringWithFormat: @"vs_timer_num_%d", j];
+            NSString *imageName = [NSString stringWithFormat: @"vs_timer_num_%ld", (long)j];
 			UIImage *sourceImage = [UIImage imageFromResource:imageName];
             
 			CGSize size		= CGSizeMake(sourceImage.size.width, sourceImage.size.height/2);
